enroll-score/requirements.txt
configparser==3.5.0
et-xmlfile==1.0.1
jdcal==1.4
openpyxl==2.5.3
python-Levenshtein==0.12.0
xlrd==1.1.0
xlwt==1.3.0
